refactor: 리팩토링 + README.md 업데이트
- 상수관리
- 예외 처리
- CRA 구조
- 경로 수정
- css 수정
