+++ README.md
@@ -1,6 +1,30 @@
-# infinite-scrolling 과제입니다.
+# 하얀마인드 | infinite-scrolling
 
-김효진 / 길용성
+### 🎉 구현 기능
+- Intersection Observer을 활용한 무한 스크롤 구현
 
-- 구현 영상
+<br>
+
+### 🎉 구현 영상
 ![ezgif com-gif-maker (1)](https://user-images.githubusercontent.com/74811374/127165887-f82f6293-84e1-40a6-af02-77d82613b35b.gif)
+---
+
+
+
+### 🎉 [리팩토링](https://brash-ridge-12d.notion.site/Assignment-c710fa753ee14bef8f4d484c7efa5a96)
+
+---
+
+[상수 관리 | API 주소](https://www.notion.so/API-e16bf09265d64a7b9a132eb41674dece)
+
+[예외 처리 | try / catch ](https://www.notion.so/try-catch-8aec752895f049c281ccfbf90ea7e3dc)
+
+[CRA Structure](https://www.notion.so/CRA-Structure-e5e2d1db84dd46e88ebf1d705796dbbd)
+
+[상대경로 → 절대경로](https://www.notion.so/254cc572ab694e77aaf68fcdaf5675d4)
+
+[의미 없는 함수명 변경](https://www.notion.so/659a26f621cf40a68febef4e4449525b)
+
+[부모 element 높이 제거하기!](https://www.notion.so/element-bf9d4d41de6846f8bda644efc2e9bf09)
+
+[useCallback 의존성 배열 줄이기](https://www.notion.so/useCallback-8d2796c5d72d4104b1d7c8448c690293)
