+++ jsconfig.json
@@ -1,0 +1,4 @@
+{ 
+	"compilerOptions": { "baseUrl": "src" }, 
+	"include": [ "src" ] 
+}
