--- src/App.css
@@ -1,3 +1,0 @@
-* {
-  box-sizing: border-box;
-}
