+++ src/App.js
@@ -1,5 +1,4 @@
-import Comment from "./components/Comment";
-import "./App.css";
+import Comment from "components/comment/Comment";
 
 function App() {
   return (
