+++ src/utils/userData.js
@@ -1,0 +1,4 @@
+export const URL = `https://jsonplaceholder.typicode.com/comments?`;
+export const LIMIT = 10;
+export const ERROR_MESSAGE =
+  "데이터를 가져오는 것을 실패하였습니다. API 주소를 확인하시거나 또는 인터넷 연결이 잘 되었는지 확인해주세요.";
